requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
botocore==1.34.145
build==1.2.2.post1
cachetools==5.4.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
chromadb==1.0.11
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cryptography==45.0.3
cycler==0.12.1
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
Deprecated==1.2.18
discord==2.3.2
discord.py==2.5.2
distlib==0.3.8
distro==1.9.0
Django==5.0.4
dnspython==2.7.0
docx2txt==0.9
durationpy==0.10
emoji==2.14.1
executing==2.0.1
fastapi==0.115.9
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.14.0
filetype==1.2.0
Flask==3.0.3
flask-cors==5.0.1
Flask-Mail==0.10.0
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.6.2
fsspec==2024.9.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
gradio==5.1.0
gradio_client==1.4.0
greenlet==3.2.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
h2==3.2.0
h5py==3.11.0
hpack==3.0.0
hstspreload==2025.1.1
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.0
huggingface-hub==0.32.2
humanfriendly==10.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
keras==3.4.1
kiwisolver==1.4.5
kubernetes==32.0.1
langchain==0.3.25
langchain-core==0.3.62
langchain-google-genai==2.0.10
langchain-text-splitters==0.3.8
langdetect==1.0.9
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1
langgraph-sdk==0.1.70
langsmith==0.3.42
libclang==18.1.1
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.0.2
ml-dtypes==0.4.0
mmh3==5.1.0
MouseInfo==0.1.3
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
namex==0.0.8
narwhals==1.41.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
ollama==0.5.1
onnxruntime==1.22.0
openai==1.82.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-instrumentation==0.54b1
opentelemetry-instrumentation-asgi==0.54b1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-util-http==0.54b1
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.7
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pdf2image==1.17.0
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.1
plotly==6.1.2
pluggy==1.5.0
posthog==4.2.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydeck==0.9.1
pydub==0.25.1
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.11.3
PyMsgBox==1.0.9
PyMuPDF==1.26.0
pynput==1.7.7
pyparsing==3.1.2
pypdf==5.5.0
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
pytest==8.3.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-iso639==2025.2.18
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.12
python-oxmsg==0.0.2
python-pptx==0.6.23
pytweening==1.2.0
pytz==2024.1
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
qrcode==8.2
qtconsole==5.5.2
QtPy==2.4.1
RapidFuzz==3.13.0
referencing==0.35.1
regex==2024.4.28
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.19.0
rsa==4.9.1
ruff==0.6.9
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.13.0
seaborn==0.13.2
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==4.1.0
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
sklearn==0.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.4.7
soupsieve==2.5
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.40.0
statsmodels==0.14.2
streamlit==1.36.0
sympy==1.13.3
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
terminado==0.18.1
threadpoolctl==3.4.0
tiktoken==0.9.0
tinycss2==1.3.0
tldextract==5.1.2
tokenizers==0.21.1
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.7.0+cu118
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.52.3
typer==0.12.5
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.99
ultralytics-thop==2.0.14
unstructured==0.17.2
unstructured-client==0.35.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.32.0
virtualenv==20.26.1
virtualenvwrapper-win==1.2.7
watchdog==4.0.1
watchfiles==1.0.5
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
wrapt==1.16.0
xgboost==2.1.1
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
zstandard==0.23.0